Modules
Organizational changes made to structure of project. Uses Flask blueprint library.
