+++ templates/a_DeleteTeam.html
@@ -90,7 +90,7 @@
                 <img src="../static/css/uploads/default.jpeg">
                  {% endif %}
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
