+++ templates/a_admincheckin.html
@@ -123,7 +123,7 @@
                             <img src="../static/css/uploads/default.jpeg">
                              {% endif %}
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
