+++ templates/a_adminteamsignup.html
@@ -124,16 +124,16 @@
             </h2>
 
             <div class="user-wrapper">
-                  <a href="{{ url_for('view') }}">
+                  <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
 
-    <form action="{{ url_for('admin_teamSignup') }}"  method="POST">
+    <form action="{{ url_for('admin.admin_teamSignup') }}"  method="POST">
         <div class="wrapper">
             <div class="signup-box">
                 <label><b>Team Name:</b><label>
