+++ templates/a_searchAndEdit.html
@@ -49,7 +49,7 @@
                <td>{{row["UserEmail"]}}</td>
                <td> {{row["RoleLevel"]}}</td>
                <td>{{row["LoginPassword"]}}</td>
-<td><a href="{{ url_for('edit_TeamForm',TeamId=row['TeamId']) }}">Edit</a></td>
+<td><a href="{{ url_for('user.edit_TeamForm',TeamId=row['TeamId']) }}">Edit</a></td>
           </tr>
               {% endfor %}
           </tbody>
