+++ templates/a_searchTeamName.html
@@ -53,7 +53,7 @@
 </head>
 <body>
     {% if session['logged_in'] and session['admin'] %}
-        <form action="{{ url_for('searchTeamName') }}" method="POST">
+        <form action="{{ url_for('both.searchTeamName') }}" method="POST">
             <h3>Search for Team</h3>
             <label for="TeamName">Name of Team/Member/Contact Info:</label><br>
             <input type="text" id="TeamName" name="TeamName" ><br>
