+++ templates/a_viewTeamSelected.html
@@ -402,11 +402,11 @@
             </h2>
 
             <div class="user-wrapper">
-                 <a href="{{ url_for('view') }}">
+                 <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
