+++ templates/a_viewTeamsAll.html
@@ -252,18 +252,18 @@
         <div class="search-wrapper">
                 <span class="las la-search"></span>
                 {% if session['logged_in'] and session['admin'] %}
-                <form action="{{ url_for('searchTeamName') }}" method="POST">
+                <form action="{{ url_for('both.searchTeamName') }}" method="POST">
                 <input type="search" placeholder="Search here"  id="TeamName" name="TeamName" ><br>
 
                 </form>
                 {% endif %}
             </div>
             <div class="user-wrapper">
-                 <a href="{{ url_for('view') }}">
+                 <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
@@ -321,7 +321,7 @@
                  {% if rows %}
                 {% for row in rows %}
                 <tr>
-                    <td><a href="{{ url_for('showOneTeam',TeamId=row['TeamId']) }}" class="but">View</a></td>
+                    <td><a href="{{ url_for('admin.showOneTeam',TeamId=row['TeamId']) }}" class="but">View</a></td>
                     <td>{{ row["TeamId"] }}</td>
                     <td>{{ row["TeamName"] }}</td>
                     <td>{{ row["SponsorName"] }}</td>
