+++ templates/dash.html
@@ -165,18 +165,18 @@
                 <span class="las la-search"></span>
 
                 {% if session['logged_in'] and session['admin'] %}
-                <form action="{{ url_for('searchTeamName') }}" method="POST">
+                <form action="{{ url_for('both.searchTeamName') }}" method="POST">
                 <input type="search" placeholder="Search here"  id="TeamName" name="TeamName" ><br>
 
                 </form>
                 {% endif %}
             </div>
             <div class="user-wrapper">
-                 <a href="{{ url_for('view') }}">
+                 <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}">
+                    <a href="{{ url_for('both.view') }}">
 
                         <h4> {{UserName}} </h4>
                     </a>
