+++ templates/login-signup.html
@@ -299,7 +299,7 @@
                 <span>Have an account? <a href="#" onclick="login()">Login</a></span>
                 <header>Sign Up</header>
             </div>
-             <form action="{{ url_for('adduser') }}"  method="POST">
+             <form action="{{ url_for('user.adduser') }}"  method="POST">
                   <form action="">
             <div class="two-forms">
                 <div class="input-box">
