+++ templates/signup.html
@@ -146,7 +146,7 @@
          {% endif %}
     {% endwith %}
 
-    <form action="{{ url_for('adduser') }}"  method="POST">
+    <form action="{{ url_for('user.adduser') }}"  method="POST">
         <div class="wrapper">
              <div class="wrapper-1">
             <div class="header">
