+++ templates/u_userteamsignup.html
@@ -115,17 +115,17 @@
             </h2>
 
             <div class="user-wrapper">
-                  <a href="{{ url_for('view') }}">
+                  <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
 
 
-    <form action="{{ url_for('user_teamSignup') }}"  method="POST">
+    <form action="{{ url_for('user.user_teamSignup') }}"  method="POST">
         <div class="wrapper">
             <div class="signup-box">
                 <label><b>Team Name:</b><label>
