+++ templates/u_viewTeamAll.html
@@ -263,18 +263,18 @@
         <div class="search-wrapper">
                 <span class="las la-search"></span>
                 {% if session['logged_in'] and session['user'] %}
-                <form action="{{ url_for('searchTeamName') }}" method="POST">
+                <form action="{{ url_for('both.searchTeamName') }}" method="POST">
                 <input type="search" placeholder="Search here"  id="TeamName" name="TeamName" ><br>
 
                 </form>
                 {% endif %}
             </div>
             <div class="user-wrapper">
-                  <a href="{{ url_for('view') }}">
+                  <a href="{{ url_for('both.view') }}">
                 <img src="../{{ photo }}" height="50" width="50">
                  </a>
                 <div>
-                    <a href="{{ url_for('view') }}"><h4> {{UserName}} </h4></a>
+                    <a href="{{ url_for('both.view') }}"><h4> {{UserName}} </h4></a>
                 </div>
             </div>
         </header>
@@ -307,7 +307,7 @@
 
                 {% else %}
                 Create your New Team Now!
-                <a href="{{ url_for('user_teamSignup') }}" class="but">New Team</a>
+                <a href="{{ url_for('user.user_teamSignup') }}" class="but">New Team</a>
             <br> ( route to create team form popup)
                  {% endif %}
         </div>
